+++ OKX.Api/Public/Responses/OkxPublicTradeAggregate.cs
@@ -1,52 +1,10 @@
 ﻿namespace OKX.Api.Public;
 
 /// <summary>
-/// OKX Trade
+/// OKX Trade All
 /// </summary>
-public record OkxPublicTradeAggregate
+public record OkxPublicTradeAggregate : OkxPublicTrade
 {
-    /// <summary>
-    /// Instrument Id
-    /// </summary>
-    [JsonProperty("instId")]
-    public string InstrumentId { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Trade Id
-    /// </summary>
-    [JsonProperty("tradeId")]
-    public long TradeId { get; set; }
-
-    /// <summary>
-    /// Price
-    /// </summary>
-    [JsonProperty("px")]
-    public decimal Price { get; set; }
-
-    /// <summary>
-    /// Quantity
-    /// </summary>
-    [JsonProperty("sz")]
-    public decimal Quantity { get; set; }
-
-    /// <summary>
-    /// Side
-    /// </summary>
-    [JsonProperty("side")]
-    public OkxTradeOrderSide Side { get; set; }
-
-    /// <summary>
-    /// Timestamp
-    /// </summary>
-    [JsonProperty("ts")]
-    public long Timestamp { get; set; }
-
-    /// <summary>
-    /// Time
-    /// </summary>
-    [JsonIgnore]
-    public DateTime Time => Timestamp.ConvertFromMilliseconds();
-
     /// <summary>
     /// Count
     /// </summary>